lib-dropbear: enable DEBUG_NOFORK
Upon receiving a connection dropbear will attempt to fork itself.
An unsuccessful attempt will result in the connection being dropped.
For testing purposes, a flag is available that will cause the client to
effectively "take over" the dropbear process.

Unikraft doesn't seem to like fork() as it returns "Function not
implemented" errno. I'm not sure if it's a Unikraft thing or perhaps it
happens in lib-musl.