lib-dropbear: do not call disallow_core
Presumably for safety purposes, dropbear sets the maximum size of a core
dump for its process to zero, thus disabling it.
My guess is that it's done to avoid leaks of any sort.

Unikraft, however, does not seem to implement the concept of core dumps.
Attempting to set this limit will result in a non-harmless message
saying "Unsupported resource". It doesn't break anything but I thought
that it would be nice to get rid of this confusing message.

Resource limits that are supported in Unikraft can be seen here:
https://github.com/unikraft/unikraft/blob/254bfe2748749336d077ba43153d8751d6faf1d9/lib/posix-process/deprecated.c#L327